ws/requirements.txt
Flask
Flask-CORS
flask_sqlalchemy
mysql-connector-python
pandas
werkzeug
PyJWT
pymysql
cryptography
python-dotenv